fix:firebaseの設定とloginを分割
