+++ templates/login.html
@@ -37,7 +37,7 @@
             <button type="submit" class="btn">Login</button>
 
             <div class="userAdd-link">
-                <a href="{{ url_for('userAdd') }}">ユーザー登録はこちら</a>
+                <a href="{{ url_for('auth.userAdd') }}">ユーザー登録はこちら</a>
             </div>
         </form>
 </body>
